chore: create and use common RxTouch util
